Refactor project structure to resemble FastAPI Tutorial example
* Use a project structure as described in the FastAPI Tutorial
* Allow running FastAPI from main to make it easy to debug
