+++ beanie_fastapi_demo/models/__init__.py
@@ -1,0 +1,3 @@
+from .note import Note
+
+__beanie_documents__ = [Note]
