+++ beanie_fastapi_demo/models/note.py
