+++ beanie_fastapi_demo/routes/__init__.py
@@ -1,0 +1,1 @@
+from .note import router as note_router
