--- beanie_fastapi_demo/run.py
@@ -1,3 +1,0 @@
-from app_factory import app
-
-app = app
